pyproject.toml
[project]
name = "folder2k8s"
authors = [
    { name = "Andrew Cole", email = "andrew.cole@illallangi.com" }
]
maintainers = [
    { name = "Andrew Cole", email = "andrew.cole@illallangi.com" }
]
description = "Synchronise HelmChart objects between a folder and kubernetes"
readme = "README.md"
requires-python = ">=3.10.1"
dependencies = [
    "diffsync>=2.0.1",
    "django-health-check>=3.18.3",
    "django>=5.1.3",
    "kubernetes>=31.0.0",
    "python-dotenv>=1.0.1",
]
keywords = [
    "kubernetes",
]
version = "0.0.0"
classifiers =[
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
]
license = { text = "MIT License" }
[project.urls]
Homepage = "https://github.com/illallangi/github2k8s"
Repository = "https://github.com/illallangi/github2k8s.git"
Issues = "https://github.com/illallangi/github2k8s/issues"
Changelog = "https://github.com/illallangi/github2k8s/blob/master/CHANGELOG.md"
[project.scripts]
folder2k8s = "folder2k8s:main"
[build-system]
requires = [
    "ruff>=0.6.9",
    "hatchling>=1.25.0",
]
build-backend = "hatchling.build"
[tool.deptry.per_rule_ignores]
DEP001 = [
    "yaml",
]
DEP002 = [
    "django-health-check",
]
[tool.ruff]
lint.select = ["ALL"]
lint.ignore = [
    "COM812",
    "D",
    "E501",
    "ERA001",
    "ISC001",
]
[tool.ruff.lint.pydocstyle]
convention = "pep257"
[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"
[tool.uv]
dev-dependencies = [
    "commitizen>=3.29.1",
    "deptry>=0.20.0",
    "hatchling>=1.25.0",
    "pre-commit>=3.8.0",
    "ruff>=0.6.9",
]
[tool.hatch.build.targets.wheel]
packages = ["src/folder2k8s"]
exclude = [
    "Makefile",
    "uv.lock",
]
[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "pep621"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version",
    "src/folder2k8s/__version__.py:__version__",
]
pre_bump_hooks = [
  "uv sync --no-dev"
]
update_changelog_on_bump = true
major_version_zero = true
